Add mcts based search
Uses net to decide on a game length, and tries to make it that long.
Good at ensuring the game gets created in the presence max policy causing move cycles, not very effective at ensuring good game play.
cpuct and fpu values are arbitrary Probably could do with a PST imeplementation, but without a strategy for tuning choosing good values will be a stretch.